Finish Basic View Structure
