+++ app/src/main/java/com/shinplest/mobiletermproject/main/MainActivity.java
@@ -2,19 +2,11 @@ package com.shinplest.mobiletermproject.main;
 
 import android.os.Bundle;
 
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentStatePagerAdapter;
 import androidx.viewpager.widget.ViewPager;
 
 import com.google.android.material.tabs.TabLayout;
 import com.shinplest.mobiletermproject.BaseActivity;
 import com.shinplest.mobiletermproject.R;
-import com.shinplest.mobiletermproject.dashboard.DashBoardFragment;
-import com.shinplest.mobiletermproject.map.MapFragment;
-import com.shinplest.mobiletermproject.record.RecordFragment;
-
-import org.jetbrains.annotations.NotNull;
 
 public class MainActivity extends BaseActivity {
     private ViewPager viewPager;
@@ -23,35 +15,17 @@ public class MainActivity extends BaseActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        viewPager = findViewById(R.id.vp_main);
-        viewPager.setAdapter(new pagerAdapter(getSupportFragmentManager()));
-        viewPager.setCurrentItem(0);
-        TabLayout tabLayout = findViewById(R.id.tab_layout);
-        tabLayout.setupWithViewPager(viewPager);
+        setViewPager();
     }
 
-    private class pagerAdapter extends FragmentStatePagerAdapter {
-        public pagerAdapter(FragmentManager fm) {
-            super(fm);
-        }
-
-        @NotNull
-        @Override
-        public Fragment getItem(int position) {
-            switch (position) {
-                case 0:
-                    return new RecordFragment();
-                case 1:
-                    return new MapFragment();
-                case 2:
-                    return new DashBoardFragment();
-            }
-            return null;
-        }
-
-        @Override
-        public int getCount() {
-            return 3;
-        }
+    private void setViewPager(){
+        viewPager = findViewById(R.id.vp_main);
+        viewPager.setAdapter(new PagerAdapter(getSupportFragmentManager()));
+        viewPager.setCurrentItem(1);
+        TabLayout tabLayout = findViewById(R.id.tab_layout);
+        tabLayout.setupWithViewPager(viewPager);
+        tabLayout.getTabAt(0).setText(getString(R.string.record));
+        tabLayout.getTabAt(1).setText(getString(R.string.map));
+        tabLayout.getTabAt(2).setText(getString(R.string.dashboard));
     }
 }
