+++ app/src/main/java/com/shinplest/mobiletermproject/main/PagerAdapter.java
@@ -1,0 +1,34 @@
+package com.shinplest.mobiletermproject.main;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentStatePagerAdapter;
+
+import com.shinplest.mobiletermproject.dashboard.DashBoardFragment;
+import com.shinplest.mobiletermproject.map.MapFragment;
+import com.shinplest.mobiletermproject.record.RecordFragment;
+
+class PagerAdapter extends FragmentStatePagerAdapter {
+    public PagerAdapter(@NonNull FragmentManager fm) {
+        super(fm, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT);
+    }
+
+    @Override
+    public Fragment getItem(int position) {
+        switch (position) {
+            case 0:
+                return new RecordFragment();
+            case 1:
+                return new MapFragment();
+            case 2:
+                return new DashBoardFragment();
+        }
+        return new MapFragment();
+    }
+
+    @Override
+    public int getCount() {
+        return 3;
+    }
+}
