+++ app/src/main/res/layout/record_fragment.xml
@@ -2,7 +2,9 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@color/white"
     android:orientation="vertical">
+
     <TextView
         android:id="@+id/brightnessText"
         android:layout_width="wrap_content"
