+++ app/src/main/res/values/colors.xml
@@ -1,6 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#6200EE</color>
-    <color name="colorPrimaryDark">#3700B3</color>
-    <color name="colorAccent">#03DAC5</color>
+    <color name="colorPrimary">#000000</color>
+    <color name="colorPrimaryDark">#69D294</color>
+    <color name="colorAccent">#000000</color>
+
+    <!--기본색상-->
+    <color name="blue">#053464</color>
+    <color name="green">#056839</color>
+    <color name="red">#7C111B</color>
+    <color name="gray">#ebebeb</color>
+    <color name="black">#000000</color>
+    <color name="white">#ffffff</color>
 </resources>
